pyproject.toml
[project]
name = "strands-agents-starter"
version = "0.1.0"
description = "Starter: Strands Agents + local LLM (DDD structure, CLI, workflow)"
authors = [
    { name = "contail", email = "qkrtkdwls2121@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "typer[all]>=0.12.0",
    "python-dotenv>=1.0.1",
    "strands-agents>=1.8.0"
]

[project.scripts]
strands-agents = "strands_agents_starter.api.cli:app"

[tool.poetry]
packages = [{include = "strands_agents_starter", from = "src"}]

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 100
target-version = "py313"
select = ["E", "F", "I", "UP"]
ignore = []

[tool.mypy]
python_version = "3.13"
strict = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
no_implicit_optional = true
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = true
disallow_untyped_defs = true
ignore_missing_imports = true


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
